+++ src/demo/fission.c
@@ -161,7 +161,7 @@ int fission_demo(struct notcurses* nc, uint64_t startns){
             ncplane_resize_simple(n, newy, newx);
             continue;
           }
-          nccell c = CELL_TRIVIAL_INITIALIZER;
+          nccell c = NCCELL_TRIVIAL_INITIALIZER;
           uint16_t smask;
           uint64_t channels;
           char* egc = ncplane_at_yx(stdn, usey, usex, &smask, &channels);