+++ src/lib/menu.c
@@ -302,7 +302,7 @@ write_header(ncmenu* ncm){
   if(ncplane_cursor_move_yx(ncm->ncp, ypos, 0)){
     return -1;
   }
-  nccell c = CELL_INITIALIZER(' ', 0, ncm->headerchannels);
+  nccell c = NCCELL_INITIALIZER(' ', 0, ncm->headerchannels);
   ncplane_set_styles(ncm->ncp, 0);
   if(ncplane_putc(ncm->ncp, &c) < 0){
     return -1;
@@ -335,7 +335,7 @@ write_header(ncmenu* ncm){
         return -1;
       }
       if(ncm->sections[i].shortcut_offset >= 0){
-        nccell cl = CELL_TRIVIAL_INITIALIZER;
+        nccell cl = NCCELL_TRIVIAL_INITIALIZER;
         if(ncplane_at_yx_cell(ncm->ncp, ypos, xoff + ncm->sections[i].shortcut_offset, &cl) < 0){
           return -1;
         }
@@ -420,7 +420,7 @@ ncmenu* ncmenu_create(ncplane* n, const ncmenu_options* opts){
           ret->sectionchannels = opts->sectionchannels;
           ret->disablechannels = ret->sectionchannels;
           ncchannels_set_fg_rgb(&ret->disablechannels, 0xdddddd);
-          nccell c = CELL_TRIVIAL_INITIALIZER;
+          nccell c = NCCELL_TRIVIAL_INITIALIZER;
           nccell_set_fg_alpha(&c, NCALPHA_TRANSPARENT);
           nccell_set_bg_alpha(&c, NCALPHA_TRANSPARENT);
           ncplane_set_base_cell(ret->ncp, &c);
@@ -519,7 +519,7 @@ int ncmenu_unroll(ncmenu* n, int sectionidx){
         }
       }
       if(sec->items[i].shortcut_offset >= 0){
-        nccell cl = CELL_TRIVIAL_INITIALIZER;
+        nccell cl = NCCELL_TRIVIAL_INITIALIZER;
         if(ncplane_at_yx_cell(n->ncp, ypos, xpos + 1 + sec->items[i].shortcut_offset, &cl) < 0){
           return -1;
         }