+++ src/tests/wide.cpp
@@ -88,7 +88,7 @@ TEST_CASE("Wide") {
     int x = 0;
     for(auto & tcell : tcells){
       CHECK(0 == ncplane_cursor_move_yx(n_, 0, x));
-      nccell testcell = CELL_TRIVIAL_INITIALIZER;
+      nccell testcell = NCCELL_TRIVIAL_INITIALIZER;
       CHECK(0 < ncplane_at_cursor_cell(n_, &testcell));
       CHECK(!strcmp(nccell_extended_gcluster(n_, &tcell), nccell_extended_gcluster(n_, &testcell)));
       CHECK(0 == testcell.stylemask);
@@ -117,7 +117,7 @@ TEST_CASE("Wide") {
     CHECK(1 == y);
     CHECK(2 == x);
     CHECK(0 < ncplane_putegc_yx(n_, 0, 0, w, nullptr));
-    nccell c = CELL_TRIVIAL_INITIALIZER;
+    nccell c = NCCELL_TRIVIAL_INITIALIZER;
     CHECK(0 < nccell_load(n_, &c, w));
     CHECK(0 < ncplane_putc_yx(n_, 1, 0, &c));
     nccell_release(n_, &c);
@@ -147,7 +147,7 @@ TEST_CASE("Wide") {
     ncplane_cursor_yx(n_, &y, &x);
     CHECK(0 == y);
     CHECK(1 + cols2 == x);
-    nccell c = CELL_TRIVIAL_INITIALIZER;
+    nccell c = NCCELL_TRIVIAL_INITIALIZER;
     ncplane_at_yx_cell(n_, 0, 0, &c);
     if(cols1 > 1){
       CHECK(0 == c.gcluster); // should be nothing
@@ -173,7 +173,7 @@ TEST_CASE("Wide") {
     ncplane_cursor_yx(n_, &y, &x);
     CHECK(0 == y);
     CHECK(3 == x);
-    nccell c = CELL_TRIVIAL_INITIALIZER;
+    nccell c = NCCELL_TRIVIAL_INITIALIZER;
     ncplane_at_yx_cell(n_, 0, 0, &c);
     if(ncstrwidth(wbashedl, NULL, NULL) > 1){
       CHECK(0 == c.gcluster); // should be nothing
@@ -202,7 +202,7 @@ TEST_CASE("Wide") {
     ncplane_cursor_yx(n_, &y, &x);
     CHECK(0 == y);
     CHECK(3 == x);
-    nccell c = CELL_TRIVIAL_INITIALIZER;
+    nccell c = NCCELL_TRIVIAL_INITIALIZER;
     ncplane_at_yx_cell(n_, 0, 0, &c);
     CHECK(0 == strcmp(nccell_extended_gcluster(n_, &c), SNAKE));
     ncplane_at_yx_cell(n_, 0, 1, &c);
@@ -235,7 +235,7 @@ TEST_CASE("Wide") {
     CHECK(0 == ncplane_rounded_box_sized(ncp, 0, 0, 3, 4, 0));
     CHECK(ncstrwidth(SCORPION, NULL, NULL) == ncplane_putegc_yx(ncp, 1, 1, SCORPION, nullptr));
     CHECK(0 == notcurses_render(nc_));
-    nccell c = CELL_TRIVIAL_INITIALIZER;
+    nccell c = NCCELL_TRIVIAL_INITIALIZER;
     CHECK(0 < ncplane_at_yx_cell(ncp, 1, 0, &c));
     CHECK(!strcmp(nccell_extended_gcluster(ncp, &c), "│"));
     nccell_release(ncp, &c);
@@ -255,7 +255,7 @@ TEST_CASE("Wide") {
 
   SUBCASE("RenderWides") {
     CHECK(0 <= ncplane_putstr(n_, "\u5f62\u5168"));
-    nccell c = CELL_TRIVIAL_INITIALIZER;
+    nccell c = NCCELL_TRIVIAL_INITIALIZER;
     ncplane_at_yx_cell(n_, 0, 0, &c);
     CHECK(nccell_double_wide_p(&c));
     ncplane_at_yx_cell(n_, 0, 1, &c);
@@ -290,7 +290,7 @@ TEST_CASE("Wide") {
   // If an ncplane is moved atop the right half of a wide glyph, the entire
   // glyph should be oblitrated.
   SUBCASE("PlaneStompsWideGlyph"){
-    nccell c = CELL_TRIVIAL_INITIALIZER;
+    nccell c = NCCELL_TRIVIAL_INITIALIZER;
     char* egc;
 
     // print two wide glyphs on the standard plane
@@ -412,7 +412,7 @@ TEST_CASE("Wide") {
     };
     struct ncplane* p = ncplane_create(n_, &nopts);
     REQUIRE(nullptr != p);
-    nccell c = CELL_CHAR_INITIALIZER('X');
+    nccell c = NCCELL_CHAR_INITIALIZER('X');
     CHECK(0 == ncplane_perimeter(p, &c, &c, &c, &c, &c, &c, 0));
     ncplane_set_bg_rgb8(n_, 0x20, 0x20, 0x20);
     CHECK(2 == ncplane_putegc_yx(n_, 1, 1, "六", nullptr));
@@ -431,7 +431,7 @@ TEST_CASE("Wide") {
     REQUIRE(nullptr != egc);
     CHECK(0 == strcmp(" ", egc));
     free(egc);
-    nccell cl = CELL_TRIVIAL_INITIALIZER, cr = CELL_TRIVIAL_INITIALIZER;
+    nccell cl = NCCELL_TRIVIAL_INITIALIZER, cr = NCCELL_TRIVIAL_INITIALIZER;
     CHECK(3 == ncplane_at_yx_cell(n_, 1, 1, &cl));
     CHECK(0 == strcmp("六", nccell_extended_gcluster(n_, &cl)));
     CHECK(0 == ncplane_at_yx_cell(n_, 1, 2, &cr));
@@ -915,7 +915,7 @@ TEST_CASE("Wide") {
   // the NUL backstop for long inlined UTF8).
   // U+1F427 PINCHED FINGERS → UTF8(f0 9f a4 8c)
   SUBCASE("ItalicEmoji") {
-    nccell c = CELL_TRIVIAL_INITIALIZER;
+    nccell c = NCCELL_TRIVIAL_INITIALIZER;
     nccell_load(n_, &c, "\U0001F427");
     CHECK(0xa7909ff0 == htole(c.gcluster));
     nccell_on_styles(&c, NCSTYLE_ITALIC);
@@ -934,7 +934,7 @@ TEST_CASE("Wide") {
   }
 
   SUBCASE("StyleMaxEmoji") {
-    nccell c = CELL_TRIVIAL_INITIALIZER;
+    nccell c = NCCELL_TRIVIAL_INITIALIZER;
     nccell_load(n_, &c, "\U0001F427");
     CHECK(0xa7909ff0 == htole(c.gcluster));
     nccell_on_styles(&c, NCSTYLE_MASK);
@@ -992,7 +992,7 @@ TEST_CASE("Wide") {
   // fill the screen with un-inlineable EGCs
 #ifndef __APPLE__  // FIXME
   SUBCASE("OfflineEGCs") {
-    nccell c = CELL_TRIVIAL_INITIALIZER;
+    nccell c = NCCELL_TRIVIAL_INITIALIZER;
     const char egc[] = "\U0001F471\u200D\u2640"; // all one EGC
     CHECK(0 < nccell_load(n_, &c, egc));
     ncplane_set_scrolling(n_, true);